Change Vector2 to use Long

+++ src/main/scala/lars/game/engine/math/Vector2.scala
@@ -1,6 +1,6 @@
 package lars.game.engine.math
 
-case class Vector2(x: Int, y: Int) {
+case class Vector2(x: Long, y: Long) {
   def +(vec: Vector2) : Vector2 = new Vector2(x + vec.x, y + vec.y)
   def -(vec: Vector2) : Vector2 = new Vector2(x - vec.x, y + vec.y)
   def ==(vec: Vector2) : Boolean = (x == vec.x && y == vec.y)
